tests/wpt/tests/css/css-sizing/min-width-max-width-precedence.html

<!DOCTYPE html>
<title>CSS Test: check precedence between min-width and max-width</title>
<link rel="help" href="https://www.w3.org/TR/CSS21/visudet.html#min-max-widths" />
<meta name="assert" content="Test that the used value of 'width' is resolved properly." />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
.tooltip {
  height: 68px;
  position: relative;
  max-width: 200px;
  min-width: 260px;
  background-color: red;
}

#outer {
  background-color: blue;
  position: absolute;
  height: 136px;
}
</style>

<div id="outer">
  <div id="inner" class="tooltip" role="tooltip"></div>
</div>
<script>
test(() => {
  assert_equals(getComputedStyle(document.querySelector("#outer")).width, '260px');
}, "When used value is resolved, min-width should win over max-width");
</script>
